mcp: move built-in tools to light module to cut per-call import cost
